requirements.txt
joblib==0.13.2
numpy==1.16.3
pandas==0.24.2
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
xgboost==0.90
absl-py==0.7.1
astor==0.8.0
gast==0.2.2
google-pasta==0.1.6
grpcio==1.21.1
h5py==2.9.0
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
numpy==1.16.3
pandas==0.24.2
protobuf==3.7.1
python-dateutil==2.8.0
pytz==2019.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tb-nightly==1.14.0a20190301
tensorflow==2.0.0a0
termcolor==1.1.0
tf-estimator-nightly==1.14.0.dev2019030115
Werkzeug==0.15.4
xgboost==0.90
